Autoauto language plugin v0.0.60

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auxilary/dsls/autoauto/compiler/functionloader/delete-unused-method-classes.js
@@ -12,6 +12,7 @@ var robotFunctionsDirectory = path.join(rootDirectory, "TeamCode/src/main/java/o
 module.exports = function(methodClassNames) {
     var javaFileNames = fs.readdirSync(robotFunctionsDirectory);
     for(var i = 0; i < javaFileNames.length; i++) {
+        if(!javaFileNames[i].endsWith(".java")) continue;
         var javaClassName = /(\w+)(\.java)?$/.exec(javaFileNames[i])[1];
         if(!methodClassNames.includes(javaClassName)) {
             fs.unlinkSync(path.join(robotFunctionsDirectory, javaClassName + ".java"));
